api places

+++ src/main/java/com/nlu/mainguyen/travelserviceapi/services/PlacesService.java
@@ -1,10 +1,13 @@
 package com.nlu.mainguyen.travelserviceapi.services;
 
+import java.net.http.HttpResponse.ResponseInfo;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
 import com.nlu.mainguyen.travelserviceapi.entities.Places;
+import com.nlu.mainguyen.travelserviceapi.model.ResponseDTO;
+import com.nlu.mainguyen.travelserviceapi.model.ResponseInfoDTO;
 import com.nlu.mainguyen.travelserviceapi.repositories.PlacesRepository;
 
 @Service
@@ -20,26 +23,44 @@ public class PlacesService {
         return this.repository.findAll();
     }
 
-    public Places create(Places input) {
-        return this.repository.save(input);
+    public Iterable<Places> findByName(String name) {
+        return this.repository.findByName(name);
     }
 
-    public Places getById(Long id) {
+    public ResponseInfoDTO create(Places input) {
+        Places result = this.repository.save(input);
+        return new ResponseInfoDTO(1, "", result);
+    }
+
+    public ResponseInfoDTO getById(Long id) {
         Optional<Places> items = this.repository.findById(id);
         if (items.isPresent()) {
-            return items.get();
+            Places result = items.get();
+            return new ResponseInfoDTO(1, "", result);
         }
-        return null;
+        return new ResponseInfoDTO(1, "", null);
     }
 
-    public void update(Places input) {
-        this.repository.save(input);
+    public ResponseDTO update(Long id, Places input) {
+        boolean exists = this.repository.existsById(id);
+        if (exists) {
+            this.repository.save(input);
+            return new ResponseDTO(1, "Thành công");
+        } else {
+            return new ResponseDTO(2, "Khong ton tai");
+        }
     }
 
-    public void deleteByID(Long id) {
-        this.repository.deleteById(id);
+    public boolean deleteByID(Long id) {
+        boolean exists = this.repository.existsById(id);
+        if (exists) {
+            this.repository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     // Lấy danh sách con theo submenuid
-    
+
 }
